handle error

+++ 2-do_deploy_web_static.py
@@ -6,7 +6,6 @@ from fabric.api import env, put, run
 from os.path import exists
 env.user = "ubuntu"
 env.hosts = ['100.27.0.202', '54.237.42.237']
-env.keys = "~/.ssh/id_rsa"
 
 def do_deploy(archive_path):
     """deploy on server web-01 and web-02"""
@@ -17,15 +16,29 @@ def do_deploy(archive_path):
         achive_file = archive_path.split("/")[-1]
         achive_file_name = achive_file.split(".")[0]
         path = "/data/web_static/releases/"
-        put(local_path=archive_path, remote_path="/tmp/")
-        run(f"mkdir -p /data/web_static/releases/{achive_file_name}")
-        run(f"tar -xzf /tmp/{achive_file} -C \
+        result = put(local_path=archive_path, remote_path="/tmp/")
+        if result.failed:
+            return False
+        result = run(f"mkdir -p /data/web_static/releases/{achive_file_name}")
+        if result.failed:
+            return False
+        result = run(f"tar -xzf /tmp/{achive_file} -C \
             {path}{achive_file_name}/")
-        run(f"mv {path}{achive_file_name}/web_static/* \
+        if result.failed:
+            return False
+        result = run(f"mv {path}{achive_file_name}/web_static/* \
             {path}{achive_file_name}/")
-        run(f"rm -rf {path}{achive_file_name}/web_static")
-        run(f"rm -rf /data/web_static/current")
-        run(f"ln -s {path}{achive_file_name}/ /data/web_static/current")
+        if result.failed:
+            return False
+        result = run(f"rm -rf {path}{achive_file_name}/web_static")
+        if result.failed:
+            return False
+        result = run(f"rm -rf /data/web_static/current")
+        if result.failed:
+            return False
+        result = run(f"ln -s {path}{achive_file_name}/ /data/web_static/current")
+        if result.failed:
+            return False
         print("New version deployed!")
     except:
         return False
